crypto_bot4.5/backend/src/api/routers/strategy.py
# strategy endpoints

from fastapi import APIRouter, HTTPException
from datetime import datetime
import sys
import os

# Add src to Python path
sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..'))

from infrastructure.exchanges.factory import ExchangeFactory, ExchangeType, Environment
from strategies.base import MarketData, PortfolioData

router = APIRouter()

# Global exchange instance
_exchange = None

async def get_exchange():
    """Get or create exchange instance."""
    global _exchange
    if _exchange is None:
        factory = ExchangeFactory()
        _exchange = factory.create_exchange(ExchangeType.COINBASE_SANDBOX, Environment.SANDBOX)
    return _exchange

async def get_portfolio_data():
    """Get real portfolio data using existing exchange methods."""
    try:
        exchange = await get_exchange()
        
        # Get real balance
        balances_raw = await exchange.get_balance()
        portfolio_balances = {}
        total_value_usd = 0
        
        # Parse balance data
        for currency, balance_info in balances_raw.items():
            if isinstance(balance_info, dict):
                available = float(balance_info.get('free', 0))
            else:
                available = float(balance_info)
            
            if available > 0:
                portfolio_balances[currency] = available
        
        # Get current prices for portfolio calculation
        current_prices = {}
        current_values = {}
        
        # Get BTC and ETH prices if we have them
        if 'BTC' in portfolio_balances:
            try:
                btc_ticker = await exchange.get_ticker('BTC/USD')
                btc_price = float(btc_ticker['last'])
                current_prices['BTC'] = btc_price
                current_values['BTC'] = portfolio_balances['BTC'] * btc_price
                total_value_usd += current_values['BTC']
            except Exception as e:
                print(f"Error getting BTC price: {e}")
                current_prices['BTC'] = 50000  # Fallback
                current_values['BTC'] = portfolio_balances['BTC'] * 50000
                total_value_usd += current_values['BTC']
        
        if 'ETH' in portfolio_balances:
            try:
                eth_ticker = await exchange.get_ticker('ETH/USD')
                eth_price = float(eth_ticker['last'])
                current_prices['ETH'] = eth_price
                current_values['ETH'] = portfolio_balances['ETH'] * eth_price
                total_value_usd += current_values['ETH']
            except Exception as e:
                print(f"Error getting ETH price: {e}")
                current_prices['ETH'] = 3000  # Fallback
                current_values['ETH'] = portfolio_balances['ETH'] * 3000
                total_value_usd += current_values['ETH']
        
        # Handle USD/USDT/USDC (assume 1:1 with USD)
        for currency in ['USD', 'USDT', 'USDC']:
            if currency in portfolio_balances:
                current_prices[currency] = 1.0
                current_values[currency] = portfolio_balances[currency]
                total_value_usd += current_values[currency]
        
        return {
            'portfolio_balances': portfolio_balances,
            'current_prices': current_prices,
            'current_values': current_values,
            'total_value_usd': total_value_usd
        }
        
    except Exception as e:
        print(f"Error getting portfolio data: {e}")
        # Return fallback data if exchange fails
        return {
            'portfolio_balances': {'USD': 100.0, 'USDC': 100.0, 'BTC': 0.5, 'ETH': 1.0},
            'current_prices': {'USD': 1.0, 'USDC': 1.0, 'BTC': 50000, 'ETH': 3000},
            'current_values': {'USD': 100.0, 'USDC': 100.0, 'BTC': 25000.0, 'ETH': 3000.0},
            'total_value_usd': 28200.0
        }

# Fixed starting portfolio value for consistent demo (never changes)
# Based on your real portfolio at snapshot time: 
# 0.5 BTC @ $104,365 + 1.0 ETH @ $2,497 + 100 USD + 100 USDC = $54,879
FIXED_START_VALUE = (0.5 * 104365) + (1.0 * 2497) + 100 + 100  # = $54,879

@router.get("/balance")
async def get_balance():
    """Get current portfolio balance with real exchange data."""
    try:
        # Get real portfolio data
        portfolio_data = await get_portfolio_data()
        
        # Calculate P&L against fixed starting value
        current_total = portfolio_data['total_value_usd']
        pnl = current_total - FIXED_START_VALUE
        pnl_percentage = (pnl / FIXED_START_VALUE) * 100 if FIXED_START_VALUE > 0 else 0
        
        # Create starting portfolio based on current real holdings
        start_portfolio = {
            'USD': 100.0,
            'USDC': 100.0, 
            'BTC': 0.5,
            'ETH': 1.0
        }
        
        # Create starting prices snapshot (fixed forever)
        start_prices = {
            'USD': 1.0,
            'USDC': 1.0,
            'BTC': 104365.0,  # Snapshot price
            'ETH': 2497.0     # Snapshot price
        }
        
        return {
            "start_portfolio": start_portfolio,
            "start_prices": start_prices,
            "start_total_value": FIXED_START_VALUE,
            "current_holdings": portfolio_data['portfolio_balances'],
            "current_prices": portfolio_data['current_prices'],
            "current_values": portfolio_data['current_values'],
            "current_total_value": current_total,
            "pnl": pnl,
            "pnl_percentage": pnl_percentage,
            "last_update": datetime.now().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.post("/refresh")
async def refresh_balance():
    """Refresh balance data (same as GET /balance but as POST for explicit refresh)."""
    return await get_balance()
