crypto_bot4.5/backend/requirements_simple.txt
# Core dependencies
openai
requests
keywordsai-tracing
python-dotenv

# OpenTelemetry instrumentation packages (matching working sample)
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
opentelemetry-instrumentation>=0.36b0
opentelemetry-instrumentation-urllib3>=0.36b0
opentelemetry-instrumentation-requests>=0.36b0
opentelemetry-instrumentation-openai>=0.36b0
opentelemetry-exporter-otlp>=1.15.0

# Trading dependencies
ccxt==4.1.77
coinbase-advanced-py==1.8.2
pandas==2.1.4
numpy==1.24.4
aiohttp==3.12.9
PyYAML==6.0.1 